Remove duplication of TestMain and cleanup config modding
